move TextResizer constants, substitute 'as const'
