+++ src/app/components/MainContent.tsx
@@ -1,7 +1,7 @@
 import { HTMLDivElement } from '@openstax/types/lib.dom';
 import React from 'react';
 import styled from 'styled-components/macro';
-import { TextResizerValue, textResizerValueMap } from '../content/components/constants';
+import { TextResizerValue, textResizerValueMap } from '../content/constants';
 import { MAIN_CONTENT_ID } from '../context/constants';
 import { Consumer } from '../context/SkipToContent';
 import { mergeRefs } from '../utils';
