+++ src/app/content/components/Page.spec.tsx
@@ -33,7 +33,7 @@ import * as searchUtils from '../search/utils';
 import * as select from '../selectors';
 import { PageReferenceError, PageReferenceMap } from '../types';
 import { formatBookData } from '../utils';
-import { textResizerDefaultValue } from './constants';
+import { textResizerDefaultValue } from '../constants';
 import ConnectedPage, { PageComponent } from './Page';
 import PageNotFound from './Page/PageNotFound';
 import allImagesLoaded from './utils/allImagesLoaded';
