+++ src/app/content/components/PageWithMockedHighlighter.spec.tsx
@@ -19,17 +19,17 @@ import { scrollTarget } from '../../navigation/selectors';
 import { AppServices, MiddlewareAPI, Store } from '../../types';
 import { assertDocument, assertWindow } from '../../utils';
 import { receivePage } from '../actions';
+import { textResizerDefaultValue } from '../constants';
 import {
   clearFocusedHighlight,
   focusHighlight,
   receiveDeleteHighlight,
   receiveHighlights,
-  requestDeleteHighlight,
+  requestDeleteHighlight
 } from '../highlights/actions';
 import * as requestDeleteHighlightHook from '../highlights/hooks/requestDeleteHighlight';
 import { initialState } from '../reducer';
 import { formatBookData } from '../utils';
-import { textResizerDefaultValue } from './constants';
 import ConnectedPage from './Page';
 import allImagesLoaded from './utils/allImagesLoaded';
