+++ src/app/content/components/constants.ts
@@ -49,12 +49,3 @@ export const maxContentGutter = 6;
 export const contentWrapperMaxWidth = contentTextWidth + sidebarDesktopWidth + verticalNavbarMaxWidth;
 
 export const defaultTheme = 'blue' as BookWithOSWebData['theme'];
-
-export const textResizerValues = [-2, -1, 0, 1, 2, 3] as const;
-export type TextResizerValue = typeof textResizerValues[number];
-const textResizerScales = [0.75, 0.9, 1, 1.25, 1.5, 2] as const;
-export const textResizerValueMap = new Map(textResizerValues.map((v, i) => [v, textResizerScales[i]]));
-export const textResizerMinValue = textResizerValues[0];
-export const textResizerMaxValue = textResizerValues[textResizerValues.length - 1];
-export const textResizerDefaultValue = textResizerValues[2];
-export const textResizerStorageKey = 'textSize';
